fixed regPage
